rename files and class names
